Fixed remove button causing the whole window to close
Remove button was causing the whole window to close
